#Not Login 을 잡기위한  System Log  의 피눈물나는 노력

+++ _3Course/lib/MT_CRingBuffer_lib/MT_CRingBuffer_lib.h
@@ -3,7 +3,7 @@
 #include <Windows.h>   
 #define WIN32_LEAN_AND_MEAN             // 거의 사용되지 않는 내용을 Windows 헤더에서 제외합니다.
 
-using ringBufferSize = int;
+using ringBufferSize = INT64;
 
 
 class CRingBuffer
